(PC-5563) move: parseDescription in libs/parsers
